renamed factory to factories
